Reinitialize arm target when teleop starts
For some reason, when the robot code starts, the robot cannot read the
current arm position accurately... x and y are fine, but z is always 0.5
meters underground.  So fix is to use the first opportunity after robot
code starts to reinitialize the arm target to the arm's current
position.

- Justin C