arm: Reset shoulder PID in auto and teleop
Reading sensor position at program startup is problematic as we seem to
get bad data from the sensor.  This causes the arm to move unexpectedly.

Try reseting should PID setpoint to current sensor angle some time after
startup when the sensor has time to flush the bad data.  Auto and teleop
init are the last chance to do so.

- Justin C